\name{easyPostgreSQL-package}
\alias{easyPostgreSQL-package}
\alias{easyPostgreSQL}
\docType{package}
\title{
\packageTitle{easyPostgreSQL}
}
\description{
\packageDescription{easyPostgreSQL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{easyPostgreSQL}
\packageIndices{easyPostgreSQL}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{easyPostgreSQL}

Maintainer: \packageMaintainer{easyPostgreSQL}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
