% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShape.R
\name{getShape}
\alias{getShape}
\title{Import a shapefile from a postgreSQL database}
\usage{
getShape(con, vecTable)
}
\arguments{
\item{con}{An object of class "pgConnect" generated by pgConnect()}

\item{vecTable}{Name of vectorial table in the database}
}
\value{
Spatial object of class 'sp'
}
\description{
This function imports a shapefile from a postgreSQL database.
}
\author{
Bruno Silva
}

