% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgArea.R
\name{pgarea}
\alias{pgArea}
\alias{pgarea}
\title{Obtains the area of a POLYGON geometry}
\usage{
pgArea(con, vecTable, geom = NULL, addColumn = FALSE)
}
\arguments{
\item{con}{An object of class "pgConnect" generated by pgConnect()}

\item{vecTable}{Name of vectorial table in the database}

\item{geom}{A string with the name of a geometry type column. Only required if the table has more than 1 geometry columns}

\item{addColumn}{TRUE or FALSE. Indicates if a column with the area should be permanently added to the table in the database}
}
\value{
A data.frame with the area of each object in the original table
}
\description{
This function obtains the area of a POLYGON geometry in a postgreSQL database.
}
\author{
Bruno Silva
}

