% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgBuffer.R
\name{pgBuffer}
\alias{pgBuffer}
\title{Obtains the buffer of a geometry}
\usage{
pgBuffer(con, vecTable, geom = NULL, dist, newTable = NULL)
}
\arguments{
\item{con}{An object of class "pgConnect" generated by pgConnect()}

\item{vecTable}{Name of vectorial table in the database}

\item{geom}{A string with the name of a geometry type column. Only required if the table has more than 1 geometry columns}

\item{dist}{Integer indicating the radius. Units of radius are measured in units of the spatial reference system. Measured in the same units has}

\item{newTable}{TRUE or FALSE. Indicates if a table with the buffer should be permanently added to the database}
}
\value{
A SpatialPolygonsDataFrame object with the buffer of each object
}
\description{
This function obtains the buffer of a geometry in a postgreSQL database.
}
\author{
Bruno Silva
}

