% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgCentroid.R
\name{pgCentroid}
\alias{pgCentroid}
\title{Obtains the centroid of a geometry}
\usage{
pgCentroid(con, vecTable, geom = NULL, newTable = NULL)
}
\arguments{
\item{con}{An object of class "pgConnect" generated by pgConnect()}

\item{vecTable}{Name of vectorial table in the database}

\item{geom}{A string with the name of a geometry type column. Only required if the table has more than 1 geometry columns}

\item{newTable}{TRUE or FALSE. Indicates if a table with the buffer should be permanently added to the database}
}
\value{
A SpatialPointsDataFrame object with the centroid of each object
}
\description{
This function obtains the centroid of a geometry in a postgreSQL database.
}
\author{
Bruno Silva
}

