% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgConnect.R
\name{pgConnect}
\alias{pgConnect}
\title{Create a connection to a postgreSQL database}
\usage{
pgConnect(dbname, host = "localhost", port = "5432", user = NULL,
  password = NULL)
}
\arguments{
\item{dbname}{name}

\item{host}{host}

\item{port}{port}

\item{user}{user}

\item{password}{password}
}
\value{
An object of class 'pgConnect'. This object is used to direct commands to the database engine.
}
\description{
Connect to a postgreSQL database going through the appropriate authorization procedure
}
\author{
Bruno Silva
}

