% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendShape.R
\name{sendShape}
\alias{sendShape}
\title{Add a shapefile to a postgreSQL database}
\usage{
sendShape(con, folder, layer, pgLayer = NULL)
}
\arguments{
\item{con}{An object of class "pgConnect" generated by pgConnect()}

\item{folder}{String with the folder containing the shapefile}

\item{layer}{A string with the name of shapefile (without the .shp extension)}

\item{pgLayer}{A string with the name of the layer in the database. Optional}
}
\value{
None
}
\description{
This function exports a shapefile to a postgreSQL database.
}
\author{
Bruno Silva
}

