% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noExport_typeGeom.R
\name{typeGeom}
\alias{typeGeom}
\title{Obtains the geometry type of a vector table}
\usage{
typeGeom(con, vecTable)
}
\arguments{
\item{con}{An object of class "connectPG" generated by dbConnectPG()}

\item{vecTable}{Name of vectorial table in the database}
}
\value{
The type of the geometry as a string. For example: "LINESTRING", "POLYGON", etc.
}
\description{
This function obtains the geometry type of a vector table in a postgreSQL database.
}
\author{
Bruno Silva
}

